Introduce errors package and refactor all error handling.
This change introduces a new errors package, imported as "errors" and
intended to replace the usage of the standard library package, with
custom error handling specific to dcrwallet.  Errors created by this
package are able to wrap other errors, describe the operation(s) which
failed, and define the kind of the error for programatic error
handling.

The project has been refactored to use the new errors package
throughout.  This replaces the previous apperrors package with an
easier to use API and fewer error codes which did not require special
handling.  Legacy error handling (such as the many different manager
errors in the udb package) has also been replaced.